+++ handlers/config.go
@@ -13,12 +13,12 @@ func (handler *AWSHandler) LoadConfigForProfile(profile string, r *http.Request,
 	w http.ResponseWriter) aws.Config {
 	var cfg aws.Config
 	var err error
-	credLoader := credential.CredentialLoader{}
+	credentialLoader := credential.CredentialLoader{}
 	if handler.multiple {
-		cfg, err = credLoader.LoadDefaultConfigForProfile(profile)
+		cfg, err = credentialLoader.LoadDefaultConfigForProfile(profile)
 		handleErr(err, "AWSConfig For multiple Profile ")
 	} else {
-		cfg, err = credLoader.LoadDefaultConfig() //config.LoadDefaultConfig(context.TODO(), config.WithRegion("us-east-1"))
+		cfg, err = credentialLoader.LoadDefaultConfig()
 		handleErr(err, "Default AWSConfig")
 	}
 	respondOnError(err, w, "Couldn't read "+profile+" profile")
@@ -27,7 +27,7 @@ func (handler *AWSHandler) LoadConfigForProfile(profile string, r *http.Request,
 
 func handleErr(err error, msg string) {
 	if err != nil {
-		log.Fatal(msg+" Load Failed err = %v", err)
+		log.Println(msg+" Load Failed err = %v", err)
 	} else {
 		log.Println(msg + "loaded successfuly")
 	}
